+++ model/EntiteSQL.class.php
@@ -2,6 +2,9 @@
 
 class EntiteSQL extends SQL
 {
+    public const TYPE_COLLECTIVITE = "collectivite";
+    public const TYPE_CENTRE_DE_GESTION = "centre_de_gestion";
+
     public const ENTITE_RACINE_DENOMINATION = "Entité racine";
     public const ID_E_ENTITE_RACINE = 0;
 
@@ -63,14 +66,14 @@ class EntiteSQL extends SQL
         $info = $this->getInfo($id_e);
 
         if (
-                $info['type'] == Entite::TYPE_COLLECTIVITE ||
-                $info['type'] == Entite::TYPE_CENTRE_DE_GESTION
+                $info['type'] == self::TYPE_COLLECTIVITE ||
+                $info['type'] == self::TYPE_CENTRE_DE_GESTION
         ) {
             return $id_e;
         }
         /** @deprecated 4.0.0 - Cas Entite::TYPE_SERVICE */
         foreach ($this->getAncetre($id_e) as $ancetre) {
-            if ($ancetre['type'] == Entite::TYPE_COLLECTIVITE) {
+            if ($ancetre['type'] == self::TYPE_COLLECTIVITE) {
                 return $ancetre['id_e'];
             }
         }
@@ -139,7 +142,7 @@ class EntiteSQL extends SQL
         if ($id_e != 0 || ! $liste_collectivite || ($liste_collectivite[0] == 0)) {
             return $this->getNavigationFilleWithType(
                 $id_e,
-                [Entite::TYPE_COLLECTIVITE,Entite::TYPE_CENTRE_DE_GESTION]
+                [self::TYPE_COLLECTIVITE,self::TYPE_CENTRE_DE_GESTION]
             );
         }
         $liste_fille = [];
@@ -329,5 +332,22 @@ class EntiteSQL extends SQL
         }
 
         throw new Exception("Aucun paramètre permettant la recherche de l'entité n'a été renseigné");
+    }
+
+    public static function getAllType()
+    {
+        return [
+            self::TYPE_COLLECTIVITE => "Collectivité",
+            self::TYPE_CENTRE_DE_GESTION => "Centre de gestion"
+        ];
+    }
+
+    public static function getNom($type)
+    {
+        $type_nom = self::getAllType();
+        if (empty($type_nom[$type])) {
+            return $type;
+        }
+        return $type_nom[$type];
     }
 }